что - то полезное сделано, но не все

+++ src/main/java/AddEvent.java
@@ -60,12 +60,14 @@ public class AddEvent {
             ResultSet resultSet = statement.executeQuery("SELECT * from events");
             LocalDate localDate = LocalDate.now();
             LocalDate local = Events.datePicker.getValue();
-            while (resultSet.next()) {
-                id = resultSet.getInt(1);
-                Date date = resultSet.getDate(4);
-                LocalDate localDate1 = LocalDate.parse(new SimpleDateFormat("yyyy-MM-dd").format(date));
-                if (localDate.isAfter(localDate1)) {
-                    statement.executeUpdate("delete from events where event_date ='" + localDate1 + "'");
+            if (resultSet.next()) {
+                while (resultSet.next()) {
+                    id = resultSet.getInt(1);
+                    Date date = resultSet.getDate(4);
+                    LocalDate localDate1 = LocalDate.parse(new SimpleDateFormat("yyyy-MM-dd").format(date));
+                    if (localDate.isAfter(localDate1)) {
+                        statement.executeUpdate("delete from events where event_date ='" + localDate1 + "'");
+                    }
                 }
             }
         } catch (SQLException throwables) {
